#87 - refactor: 유저 Controller 단일 책임 원칙 적용
