+++ src/main/java/com/gxdxx/instagram/domain/user/api/UserCreateController.java
@@ -1,0 +1,60 @@
+package com.gxdxx.instagram.domain.user.api;
+
+import com.gxdxx.instagram.domain.user.application.UserCreateService;
+import com.gxdxx.instagram.domain.user.dto.request.UserSignUpRequest;
+import com.gxdxx.instagram.domain.user.dto.response.UserSignUpResponse;
+import com.gxdxx.instagram.global.common.dto.response.ErrorResponse;
+import com.gxdxx.instagram.global.error.InvalidRequestException;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.MediaType;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.multipart.MultipartFile;
+
+@Tag(name = "users", description = "회원 API")
+@RequiredArgsConstructor
+@RequestMapping("/api/users")
+@RestController
+public class UserCreateController {
+
+    private final UserCreateService userCreateService;
+
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "회원가입 성공",
+                    content = @Content(schema = @Schema(implementation = UserSignUpResponse.class))),
+            @ApiResponse(responseCode = "400", description = "잘못된 요청",
+                    content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
+    })
+    @Operation(summary = "회원가입 메소드", description = "회원가입 메소드입니다.")
+    @PostMapping(value = "/signup", consumes = MediaType.MULTIPART_FORM_DATA_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
+    public UserSignUpResponse createUser(
+            @Valid UserSignUpRequest request,
+            BindingResult bindingResult
+    ) {
+        validateRequest(bindingResult);
+        validateProfileImage(request.profileImage());
+        return userCreateService.createUser(request);
+    }
+
+    private void validateRequest(BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            throw new InvalidRequestException();
+        }
+    }
+
+    private void validateProfileImage(MultipartFile profileImage) {
+        if (profileImage.isEmpty()) {
+            throw new InvalidRequestException();
+        }
+    }
+
+}
