+++ src/main/java/com/gxdxx/instagram/domain/user/api/UserDeleteController.java
@@ -1,0 +1,56 @@
+package com.gxdxx.instagram.domain.user.api;
+
+import com.gxdxx.instagram.domain.user.application.UserDeleteService;
+import com.gxdxx.instagram.domain.user.dto.request.UserDeleteRequest;
+import com.gxdxx.instagram.global.common.dto.response.ErrorResponse;
+import com.gxdxx.instagram.global.common.dto.response.SuccessResponse;
+import com.gxdxx.instagram.global.error.InvalidRequestException;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.MediaType;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.security.Principal;
+
+@Tag(name = "users", description = "회원 API")
+@RequiredArgsConstructor
+@RequestMapping("/api/users")
+@RestController
+public class UserDeleteController {
+
+    private final UserDeleteService userDeleteService;
+
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "204", description = "회원탈퇴 성공",
+                    content = @Content(schema = @Schema(implementation = SuccessResponse.class))),
+            @ApiResponse(responseCode = "400", description = "잘못된 요청",
+                    content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
+    })
+    @Operation(summary = "회원탈퇴 메소드", description = "회원탈퇴 메소드입니다.")
+    @DeleteMapping(value = "/{id}", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
+    public SuccessResponse deleteUser(
+            @Valid UserDeleteRequest request,
+            BindingResult bindingResult,
+            Principal principal
+    ) {
+        validateRequest(bindingResult);
+        return userDeleteService.deleteUser(request.password(), principal.getName());
+    }
+
+    private void validateRequest(BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            throw new InvalidRequestException();
+        }
+    }
+
+
+}
