+++ src/main/java/com/gxdxx/instagram/domain/user/api/UserLoginController.java
@@ -1,0 +1,55 @@
+package com.gxdxx.instagram.domain.user.api;
+
+import com.gxdxx.instagram.domain.user.application.UserLoginService;
+import com.gxdxx.instagram.domain.user.dto.request.UserLoginRequest;
+import com.gxdxx.instagram.global.common.dto.response.ErrorResponse;
+import com.gxdxx.instagram.global.common.dto.response.SuccessResponse;
+import com.gxdxx.instagram.global.error.InvalidRequestException;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.MediaType;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Tag(name = "users", description = "회원 API")
+@RequiredArgsConstructor
+@RequestMapping("/api/users")
+@RestController
+public class UserLoginController {
+
+    private final UserLoginService userLoginService;
+
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "로그인 성공",
+                    content = @Content(schema = @Schema(implementation = SuccessResponse.class))),
+            @ApiResponse(responseCode = "400", description = "잘못된 요청",
+                    content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
+    })
+    @Operation(summary = "로그인 메소드", description = "로그인 메소드입니다. 토큰이 발급됩니다.")
+    @PostMapping(value = "/login", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
+    public SuccessResponse login(
+            @RequestBody @Valid UserLoginRequest request,
+            BindingResult bindingResult,
+            HttpServletResponse response
+    ) {
+        validateRequest(bindingResult);
+        return userLoginService.login(request, response);
+    }
+
+    private void validateRequest(BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            throw new InvalidRequestException();
+        }
+    }
+
+}
